Use a 16bit integer for audio format to remove warning
On CI we get the following warning:
> sound.c:60:34: warning: passing argument 2 of 'Mix_QuerySpec' from incompatible pointer type [-Wincompatible-pointer-types]

This is because we're using a default (32-bit) integer pointer when the
function expects a pointer to a 16-bit integer.

+++ ext/ruby2d/sound.c
@@ -53,7 +53,7 @@ int R2D_GetSoundLength(R2D_Sound *snd) {
   float points = 0;
   float frames = 0;
   int frequency = 0;
-  int format = 0;
+  Uint16 format = 0;
   int channels = 0;
 
   // Populate the frequency, format and channel variables
